+++ core/editor/interface.go
@@ -52,6 +52,9 @@ type Edit interface {
 	Undo()
 	Redo()
 	Apply(entry log.Entry)
+
+	InsertLine(offset int, lines [][]rune)
+	DeleteLine(offset int, count int)
 }
 
 type Render interface {