+++ CHANGELOG.md
@@ -32,11 +32,9 @@
 - added regexp search
 - refactor - VISUAL/COMMAND/INSERT is now default
 - rename VISUAL -> NORMAL
+- add using `V`, `d`, `y`, `p` to select, copy, cut, paste lines. probably need another mode `SELECT` and edit view to support arbitrary highlighting
 
 # TODO
 
-- [experimental] remove `Header`, `Command` from `Status` - add `Status.Other map[string]any` to inject arbitrary data into `renderCh`
-
-- add using `V`, `d`, `y`, `p` to select, copy, cut, paste lines. probably need another mode `SELECT` and edit view to support arbitrary highlighting 
 
 - add compressed log: compress multiple `type` into a single `type` (`Text []rune` instead of `Rune rune`)